objdetectapi/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
Django==3.0.6
djangorestframework==3.11.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
importlib-metadata==1.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.2
numpy==1.18.4
opencv-python==4.2.0.32
pandas==1.0.3
Pillow==5.1.0
protobuf==3.11.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==1.0.1
zipp==3.1.0
